kafpubsub/requirements.txt
google-cloud-pubsub
confluent-kafka
